The first complete version.

+++ src/shiftregister.hpp
@@ -3,4 +3,4 @@
 
 #include <Arduino.h>
 
-void shift_register_send(int dataPin, int serialClockPin, int registerClockPin, size_t length, const bool *data);
+void shift_register_send(int dataPin, int serialClockPin, int registerClockPin, int length, const bool *data);
